pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-dynamic-serializer"
version = "0.1.0"
description = "Dynamically select which fields a Django REST Framework serializer returns, reducing payload size and improving database performance by avoiding unnecessary serialization of unrequested fields and relations."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

authors = [
  {name = "Jackson Alves", email = "jacksonalves541@gmail.com"},
]

keywords = ["django", "rest-framework", "serializer", "dynamic-fields", "performance"]

dependencies = [
  "django>=3.2",
  "djangorestframework>=3.0",
]

classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Django",
  "Framework :: Django :: 3.2",
  "Framework :: Django :: 4.0",
  "Framework :: Django :: 4.1",
  "Framework :: Django :: 4.2",
  "Framework :: Django :: 5.0",
  "Framework :: Django :: 5.1",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.optional-dependencies]
test = ["django-virtual-models>=0.1.4"]

[project.urls]
Homepage = "https://github.com/jackson541/django-dynamic-serializer"
Repository = "https://github.com/jackson541/django-dynamic-serializer"
Issues = "https://github.com/jackson541/django-dynamic-serializer/issues"

[tool.setuptools.packages.find]
include = ["django_dynamic_serializer*"]
